safari-600.7-branch/LayoutTests/media/media-source/media-source-multiple-initialization-segments-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
Test that a replacement initialization segment containing a track with the same codec but a different trackID succeeds.
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
Test that a replacement initialization segment containing a track with a different codec but the same trackID fails.
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(sourceended)
END OF TEST
